+++ src/styles/theme/palette.ts
