+++ src/styles/theme/theme.ts
@@ -1,5 +1,5 @@
 import typography from "./typography";
-import palette from "src/styles/palette";
+import palette from "src/styles/theme/palette";
 
 const theme = {
   palette,
